Distribute functions for validation
